requirements.txt
pandas~=2.2.2
matplotlib~=3.9.0
plotly
scikit-learn~=1.5.0
xgboost~=2.0.3
optuna~=3.6.1
# celery~=5.4.0
PyYAML~=6.0.1
validators~=0.28.3
joblib~=1.4.2
numpy~=1.26.4
flask~=3.0.3
SQLAlchemy~=2.0.30
requests~=2.32.3